makefile en cours

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: chsimon <chsimon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/07 00:44:09 by chsimon           #+#    #+#             */
-/*   Updated: 2021/11/24 19:41:05 by chsimon          ###   ########.fr       */
+/*   Updated: 2021/11/25 19:35:30 by chsimon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	unsigned long int	i;
 
 	i = 0;
 	if (*(unsigned char *)s == (unsigned char)c)
